+++ palace/linalg/strumpack.hpp
@@ -26,14 +26,15 @@ private:
 public:
   StrumpackSolverBase(MPI_Comm comm, SymbolicFactorization reorder,
                       SparseCompression compression, double lr_tol, int butterfly_l,
-                      int lossy_prec, int print);
+                      int lossy_prec, bool reorder_reuse, int print);
 
   StrumpackSolverBase(const IoData &iodata, MPI_Comm comm, int print)
     : StrumpackSolverBase(comm, iodata.solver.linear.sym_factorization,
                           iodata.solver.linear.strumpack_compression_type,
                           iodata.solver.linear.strumpack_lr_tol,
                           iodata.solver.linear.strumpack_butterfly_l,
-                          iodata.solver.linear.strumpack_lossy_precision, print)
+                          iodata.solver.linear.strumpack_lossy_precision,
+                          iodata.solver.linear.reorder_reuse, print)
   {
   }